// src/app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  STAFF: 'STAFF',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const DepartmentStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type DepartmentStatus = (typeof DepartmentStatus)[keyof typeof DepartmentStatus]


export const RoomStatus = {
  AVAILABLE: 'AVAILABLE',
  FULL: 'FULL',
  MAINTENANCE: 'MAINTENANCE'
} as const

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const PatientStatus = {
  ADMITTED: 'ADMITTED',
  DISCHARGED: 'DISCHARGED'
} as const

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus]


export const MedicineType = {
  PILL: 'PILL',
  INJECTION: 'INJECTION',
  LIQUID: 'LIQUID',
  POWDER: 'POWDER',
  CREAM: 'CREAM',
  PATCH: 'PATCH',
  DROPS: 'DROPS',
  OTHER: 'OTHER'
} as const

export type MedicineType = (typeof MedicineType)[keyof typeof MedicineType]


export const BoxStatus = {
  ACTIVE: 'ACTIVE',
  RETURNED: 'RETURNED'
} as const

export type BoxStatus = (typeof BoxStatus)[keyof typeof BoxStatus]


export const AssignmentStatus = {
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  PAUSED: 'PAUSED',
  CANCELLED: 'CANCELLED'
} as const

export type AssignmentStatus = (typeof AssignmentStatus)[keyof typeof AssignmentStatus]


export const AdministrationStatus = {
  GIVEN: 'GIVEN',
  SKIPPED: 'SKIPPED',
  DELAYED: 'DELAYED'
} as const

export type AdministrationStatus = (typeof AdministrationStatus)[keyof typeof AdministrationStatus]